refactor(frontend): split solutions into terrestrial and marine
Split `solutions` config into `marine` and `terrestrial`, so that each view layer has its own config directory.
