+++ frontend/src/config/marine/marine-view-layer.ts
@@ -7,7 +7,7 @@ import { selectableMvtLayer } from 'lib/deck/layers/selectable-mvt-layer';
 import { dataColorMap } from 'lib/deck/props/color-map';
 import { fillColor } from 'lib/deck/props/style';
 
-import { SolutionHoverDescription } from './SolutionHoverDescription';
+import { SolutionHoverDescription } from './MarineHoverDescription';
 
 function filterRange(value: boolean) {
   return value ? [1, 1] : [0, 1];
