+++ frontend/src/config/marine/styles.ts
@@ -1,0 +1,3 @@
+import { makeConfig } from 'lib/helpers';
+
+export const MARINE_STYLES = makeConfig([{ id: 'habitat', label: 'Habitat' }]);
