+++ frontend/src/config/terrestrial/colors.ts
@@ -1,5 +1,4 @@
 import { makeColorConfig } from 'lib/helpers';
-import { MarineHabitatType } from './domains';
 
 export const TERRESTRIAL_LANDUSE_COLORS = makeColorConfig({
   Bamboo: '#8cae26',
@@ -29,15 +28,4 @@ export const TERRESTRIAL_LANDUSE_COLORS = makeColorConfig({
   'Secondary Forest': '#73ba6b',
   'Swamp Forest': '#37d47e',
   'Water Body': '#6bdbe9',
-});
-
-export const MARINE_HABITAT_COLORS = makeColorConfig<MarineHabitatType>({
-  coral: '#f9b2ea',
-  coral_mangrove: '#808cf2',
-  coral_mangrove_seagrass: '#152513',
-  coral_seagrass: '#f2808c',
-  mangrove: '#80f2e6',
-  mangrove_seagrass: '#95e78b',
-  seagrass: '#f2e680',
-  other: 'rgba(128,128,128,0.2)',
 });
