+++ frontend/src/config/terrestrial/domains.ts
@@ -1,5 +1,3 @@
-import { toDictionary } from 'lib/helpers';
-
 const landUseValuesConst = [
   'Bamboo',
   'Bamboo and Fields',
@@ -62,64 +60,3 @@ export const TERRESTRIAL_LOCATION_FILTERS = [
 ] as const;
 
 export type TerrestrialLocationFilterType = (typeof TERRESTRIAL_LOCATION_FILTERS)[number]['value'];
-
-const marineHabitatsConst = [
-  {
-    value: 'coral',
-    label: 'Coral',
-  },
-  {
-    value: 'coral_seagrass',
-    label: 'Coral and Seagrass',
-  },
-  {
-    value: 'seagrass',
-    label: 'Seagrass',
-  },
-  {
-    value: 'mangrove',
-    label: 'Mangrove',
-  },
-  {
-    value: 'coral_mangrove_seagrass',
-    label: 'Coral, Mangrove and Seagrass',
-  },
-  {
-    value: 'mangrove_seagrass',
-    label: 'Mangrove and Seagrass',
-  },
-  {
-    value: 'coral_mangrove',
-    label: 'Coral and Mangrove',
-  },
-  {
-    value: 'other',
-    label: 'Buffer Zone',
-  },
-] as const;
-
-export const MARINE_HABITATS = [...marineHabitatsConst];
-export const MARINE_HABITATS_LOOKUP = toDictionary(
-  MARINE_HABITATS,
-  (x) => x.value,
-  (x) => x.label,
-);
-
-export type MarineHabitatType = (typeof MARINE_HABITATS)[number]['value'];
-
-export const MARINE_LOCATION_FILTERS = [
-  {
-    value: 'within_coral_500m',
-    label: 'Within 500m of coral',
-  },
-  {
-    value: 'within_seagrass_500m',
-    label: 'Within 500m of seagrass',
-  },
-  {
-    value: 'within_mangrove_500m',
-    label: 'Within 500m of mangrove',
-  },
-] as const;
-
-export type MarineLocationFilterType = (typeof MARINE_LOCATION_FILTERS)[number]['value'];
