+++ frontend/src/sidebar/solutions/MarineControl.tsx
@@ -1,4 +1,4 @@
-import { MarineLocationFilterType, MARINE_LOCATION_FILTERS } from 'config/solutions/domains';
+import { MarineLocationFilterType, MARINE_LOCATION_FILTERS } from 'config/marine/domains';
 import { ParamChecklist } from 'lib/controls/params/ParamChecklist';
 import { useRecoilState } from 'recoil';
 import { InputSection } from 'sidebar/ui/InputSection';
